+++ cmd/amb-sidecar/filters/handler/httpclient/httpcache.go
@@ -56,7 +56,7 @@ func (fn roundTripperFunc) RoundTrip(req *http.Request) (*http.Response, error) 
 //        violation of RFC7234)
 //  - It logs all requests+responses, and whether or not they came
 //    from the network for from the cache.
-func NewHTTPClient(logger types.Logger, maxStale time.Duration, insecure bool) *http.Client {
+func NewHTTPClient(logger types.Logger, maxStale time.Duration, insecure bool, renegotiate bool) *http.Client {
 	return &http.Client{
 		Transport: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
 			if maxStale > 0 {
@@ -84,9 +84,17 @@ func NewHTTPClient(logger types.Logger, maxStale time.Duration, insecure bool) *
 					cached = false
 					start := time.Now()
 					transport := http.DefaultTransport
-					if insecure {
+					if insecure || renegotiate {
 						// this is the definition of http.DefaultTransport,
 						// but with TLSClientConfig added.
+						renegotiationSupport := tls.RenegotiateNever
+						if renegotiate {
+							renegotiationSupport = tls.RenegotiateOnceAsClient
+						}
+						tlsConfig := &tls.Config{
+							Renegotiation:      renegotiationSupport,
+							InsecureSkipVerify: insecure,
+						}
 						transport = &http.Transport{
 							Proxy: http.ProxyFromEnvironment,
 							DialContext: (&net.Dialer{
@@ -99,10 +107,7 @@ func NewHTTPClient(logger types.Logger, maxStale time.Duration, insecure bool) *
 							TLSHandshakeTimeout:   10 * time.Second,
 							ExpectContinueTimeout: 1 * time.Second,
 							// #nosec G402
-							TLSClientConfig: &tls.Config{
-								Renegotiation:      tls.RenegotiateOnceAsClient,
-								InsecureSkipVerify: true,
-							},
+							TLSClientConfig: tlsConfig,
 						}
 					}
 					res, err := transport.RoundTrip(req)